Closes #2
modify_mastertable.py is just a script example. Specific uses depend very much on user's intention. Although we can give support and guidance on how to correctly format a mastertable, this script is not meant for general use, and thus its specific variations fall out of the scope of this repo.